refactor(frontend): remove the Today tab and relocate its live cards
Delete the Today bottom-nav tab, TodayScreen, and dead Today-feature code
now that the hero, stat tiles, and Return/invitation stacks live on the
Journal shelf. Relocate the still-live modules out of features/Today into
two durable homes — features/Return/ (ReturnStack + Metta/Return cards,
copy, hooks) and features/Invitations/ (InvitationStack, InvitationNote,
useInvitations, copy) — and repoint every importer. Collapsing the dual
mount also resolves the interim useMettaReturn/useInvitations state
divergence flagged in PR #1241's review.
