+++ frontend/src/features/Invitations/InvitationNote.tsx
