+++ frontend/src/features/Journal/JournalShelfScreen.tsx
@@ -25,8 +25,8 @@ import { Button } from '@/components/Button';
 import { EmptyState } from '@/components/feedback/EmptyState';
 import { ScreenHeader } from '@/components/layout/ScreenHeader';
 import { ScreenScaffold } from '@/components/layout/ScreenScaffold';
-import InvitationStack from '@/features/Today/InvitationStack';
-import ReturnStack from '@/features/Today/ReturnStack';
+import InvitationStack from '@/features/Invitations/InvitationStack';
+import ReturnStack from '@/features/Return/ReturnStack';
 import { useReducedMotion } from '@/hooks/useReducedMotion';
 import type { RootStackParamList } from '@/navigation/RootStack';
 import { useDerivedCurrentWeek } from '@/store/useProgramProgression';
